requirements.txt
Pillow
tqdm
requests
matplotlib
transformers
openai
allpairspy
scikit-optimize